Added code to parse and check the PDF content of the UPW Assessment pdf in Delius
